Start rewrite to Astro.build
Well, this is quite hard... Animations are broken, almost everything
needs to be rewritten as it doesn't work anymore. I need to manually
add tools such as eslint or prettier. The Sentry's adblock bypasser is
not implemented here. And many-many other things are broken here, I do
not know even how to debug this. There is no `api/og` anymore. And so
on.

I need to sleep and rest from this...